+++ src/elements/conditions/episodes/EpisodeSeasonNameConditionRule.php
@@ -1,0 +1,52 @@
+<?php
+/**
+ * @copyright Copyright (c) vnali
+ */
+
+namespace vnali\studio\elements\conditions\episodes;
+
+use Craft;
+use craft\base\conditions\BaseTextConditionRule;
+
+use craft\base\ElementInterface;
+use craft\elements\conditions\ElementConditionRuleInterface;
+use craft\elements\db\ElementQueryInterface;
+use vnali\studio\elements\db\EpisodeQuery;
+use vnali\studio\elements\Episode;
+
+class EpisodeSeasonNameConditionRule extends BaseTextConditionRule implements ElementConditionRuleInterface
+{
+    /**
+     * @inheritdoc
+     */
+    public function getLabel(): string
+    {
+        return Craft::t('studio', 'Episode Season Name');
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getExclusiveQueryParams(): array
+    {
+        return ['seasonName'];
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function modifyQuery(ElementQueryInterface $query): void
+    {
+        /** @var EpisodeQuery $query */
+        $query->seasonName($this->paramValue());
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function matchElement(ElementInterface $element): bool
+    {
+        /** @var Episode $element */
+        return $this->matchValue($element->episodeSeasonName);
+    }
+}